Add local file server
Instead of serving the index.html file as an encoded url,
we create a file server every time a sketch is open.

All this to enable users to use the loadImage function and
such.

Sadly, the node http package only works on the main process,
therefore had to do a bunch of ipc message passing.