Change hot-code reload behaviour
Now we check if there was an exception in the code-parser,
which usually indicates that the code is invalid. This is a normal
case when the user is editing the code, like removing values or
writing down variable names.

When an error occurs, we want to just do nothing: don't reload the
preview pane, not parse the new files nor send new variables to the
preview pane.