Well, that was a lot of work for nothing
The problem I was trying to solve with hot-reload is as follows:

let x = 10;
let fillColor = 1;

function draw() {
  background(255);
  fill(fillColor);
  rect(x, x, 10);
}

The parsed code becomes, with fake keys for this text:

const __AllVars = {
  x: 10,
  fillColor: 1,
  backgroundColor: 255,
  rectSize: 10,
};

let x = __AllVars[x];
let fillColor = __AllVars[fillColor];

function draw() {
  background(__AllVars[backgroundColor]);
  fill(fillColor);
  rect(x, x, __AllVars[rectSize]);
}

Changing, say, the `fillColor` in the original code doesn't have an effect on
the sketch, at all. To see the change, it'd be necessary to
reload the sketch. That makes sense as the code has a reference to the variable
itself, and not to __AllVars[fillColor].
The same behaviour is not observed when changing variables local to the scope,
as we see the updated values in the running sketch, which's the whole point of
having hot-code reloading.

I thought that this might have something to do with how P5 works, and thought
that, perhaps, things would work if the parsed code were to inline all the
variable references, as such:

const __AllVars = {
  x: 10,
  fillColor: 1,
  backgroundColor: 255,
  rectSize: 10,
};

let x = __AllVars[x];
let fillColor = __AllVars[fillColor];

function draw() {
  background(__AllVars[backgroundColor]);
  fill(__AllVars[backgroundColor]); // difference here
  rect(__AllVars[x], __AllVars[x], // and in these two references
    __AllVars[rectSize]);
}

This commit does that, however, there's an undesired side-effect: now, when
updating any variable, _all_ of them are reset to their original values, e.g.:
when changing the `fillColor`, the `x` varable is also reset to its original
value of `10` in this case, also rendering the whole things useless.

Which leads me to the now obvious conlusion: this whole hot-reload thing is only
useful in its current form when changing literal values being used directly in
the code, without the use of variables.